+++ src/zzz_od/application/coffee/coffee_app.py
@@ -27,13 +27,13 @@ from zzz_od.application.coffee.coffee_config import (
 )
 from zzz_od.application.zzz_application import ZApplication
 from zzz_od.context.zzz_context import ZContext
-from zzz_od.controller.turn_compensation import AngleTurnCompensator
 from zzz_od.game_data.compendium import Coffee
 from zzz_od.operation.back_to_normal_world import BackToNormalWorld
 from zzz_od.operation.compendium.area_patrol import AreaPatrol
 from zzz_od.operation.compendium.combat_simulation import CombatSimulation
 from zzz_od.operation.compendium.expert_challenge import ExpertChallenge
 from zzz_od.operation.transport import Transport
+from zzz_od.operation.turning.turn_compensation import AngleTurnCompensator
 from zzz_od.operation.wait_normal_world import WaitNormalWorld
